+++ id/id/src/main/java/io/piano/android/id/PianoId.kt
@@ -5,6 +5,12 @@ import android.os.Build
 import androidx.annotation.RestrictTo
 import androidx.annotation.VisibleForTesting
 import com.squareup.moshi.Moshi
+import com.squareup.moshi.Types
+import io.piano.android.consents.ConsentJsonAdapterFactory
+import io.piano.android.consents.PianoConsents
+import io.piano.android.consents.models.Consent
+import io.piano.android.consents.models.Purpose
+import io.piano.android.id.models.ConsentData
 import io.piano.android.id.models.PianoIdToken
 import io.piano.android.id.models.PianoUserInfo
 import io.piano.android.id.models.PianoUserProfile
@@ -24,7 +30,60 @@ class PianoId {
         @VisibleForTesting
         @RestrictTo(RestrictTo.Scope.LIBRARY)
         @JvmStatic
+        @Volatile
         private var client: PianoIdClient? = null
+
+        @JvmStatic
+        private fun buildClient(
+            endpoint: HttpUrl,
+            aid: String,
+            pianoConsents: PianoConsents?,
+            useCustomHost: Boolean,
+        ): PianoIdClient {
+            val userAgent = "Piano ID SDK ${BuildConfig.SDK_VERSION} (Android ${Build.VERSION.RELEASE})"
+            val moshi = Moshi.Builder()
+                .add(PianoIdJsonAdapterFactory())
+                .add(UnixTimeDateAdapter)
+                .add(ConsentJsonAdapterFactory)
+                .build()
+            val storage = pianoConsents ?: runCatching { PianoConsents.getInstance() }.getOrNull()
+            val consentsDataProvider = ConsentsDataProvider(
+                storage,
+                moshi.adapter(
+                    Types.newParameterizedType(
+                        Map::class.java,
+                        Purpose::class.java,
+                        Consent::class.java
+                    )
+                ),
+                moshi.adapter(
+                    Types.newParameterizedType(
+                        List::class.java,
+                        ConsentData::class.java
+                    )
+                )
+            )
+            val okHttpClient = OkHttpClient.Builder()
+                .addInterceptor(ConsentsInterceptor(consentsDataProvider))
+                .addInterceptor(UserAgentInterceptor(userAgent))
+                .addInterceptor(AidInterceptor(aid))
+                .addInterceptor(
+                    HttpLoggingInterceptor().setLevel(
+                        if (BuildConfig.DEBUG || isLogHttpSet()) {
+                            HttpLoggingInterceptor.Level.BODY
+                        } else {
+                            HttpLoggingInterceptor.Level.NONE
+                        }
+                    )
+                )
+                .build()
+            val retrofit = Retrofit.Builder()
+                .client(okHttpClient)
+                .baseUrl(endpoint)
+                .addConverterFactory(MoshiConverterFactory.create(moshi))
+                .build()
+            return PianoIdClient(retrofit.create(), moshi, aid, consentsDataProvider, endpoint, useCustomHost)
+        }
 
         /**
          * Initialize {@link PianoIdClient} singleton instance. It doesn't re-init it at next calls.
@@ -32,49 +91,51 @@ class PianoId {
          * @param endpoint Endpoint, which will be used. For example, {@link #ENDPOINT_PRODUCTION},
          *                 {@link #ENDPOINT_SANDBOX} or your custom endpoint
          * @param aid      Your AID
-         * @param useCustomHost a flag if a custom deployment host should be used ([endpoint] will be used as is)
+         * @param pianoConsents [PianoConsents] instance for managing user consent
+         * @param useCustomHost If true [endpoint] will be used as deployment host else (default) the host gets overridden internally.
          * @return {@link PianoIdClient} instance.
          */
         @Suppress("unused") // Public API.
         @JvmStatic
-        fun init(endpoint: String, aid: String, useCustomHost: Boolean = false): PianoIdClient = init(endpoint.toHttpUrl(), aid, useCustomHost)
+        fun init(
+            endpoint: String,
+            aid: String,
+            pianoConsents: PianoConsents? = null,
+            useCustomHost: Boolean = false,
+        ): PianoIdClient = init(endpoint.toHttpUrl(), aid, pianoConsents, useCustomHost)
 
         /**
          * Initialize {@link PianoIdClient} singleton instance. It doesn't re-init it at next calls.
          *
          * @param endpoint Endpoint, which will be used.
          * @param aid      Your AID
+         * @param pianoConsents [PianoConsents] instance for managing user consent
          * @param useCustomHost If true [endpoint] will be used as deployment host else (default) the host gets overridden internally.
          * @return {@link PianoIdClient} instance.
          */
         @Suppress("unused") // Public API.
         @JvmStatic
-        fun init(endpoint: HttpUrl, aid: String, useCustomHost: Boolean = false): PianoIdClient =
-            client ?: run {
-                val userAgent = "Piano ID SDK ${BuildConfig.SDK_VERSION} (Android ${Build.VERSION.RELEASE})"
-                val okHttpClient = OkHttpClient.Builder()
-                    .addInterceptor(UserAgentInterceptor(userAgent))
-                    .addInterceptor(
-                        HttpLoggingInterceptor().setLevel(
-                            if (BuildConfig.DEBUG || isLogHttpSet()) {
-                                HttpLoggingInterceptor.Level.BODY
-                            } else {
-                                HttpLoggingInterceptor.Level.NONE
-                            }
-                        )
-                    )
-                    .build()
-                val moshi = Moshi.Builder()
-                    .add(PianoIdJsonAdapterFactory())
-                    .add(UnixTimeDateAdapter)
-                    .build()
-                val retrofit = Retrofit.Builder()
-                    .client(okHttpClient)
-                    .baseUrl(endpoint)
-                    .addConverterFactory(MoshiConverterFactory.create(moshi))
-                    .build()
-                PianoIdClient(retrofit.create(), moshi, aid, endpoint, useCustomHost).also { client = it }
+        fun init(
+            endpoint: HttpUrl,
+            aid: String,
+            pianoConsents: PianoConsents? = null,
+            useCustomHost: Boolean = false,
+        ): PianoIdClient {
+            if (client == null) {
+                synchronized(this) {
+                    if (client == null) {
+                        client = buildClient(endpoint, aid, pianoConsents, useCustomHost)
+                    }
+                }
             }
+            return getInstance()
+        }
+
+        @JvmStatic
+        @Suppress("unused") // Public API.
+        fun getInstance(): PianoIdClient = checkNotNull(client) {
+            NOT_INITIALIZED_MSG
+        }
 
         /**
          * Gets preferences for authorization process
@@ -85,7 +146,11 @@ class PianoId {
         @Suppress("unused") // Public API.
         @Throws(IllegalStateException::class)
         @JvmStatic
-        fun signIn(): PianoIdClient.SignInContext = getClient().signIn()
+        @Deprecated(
+            "Use method directly from PianoIdClient instance, will be removed in future versions",
+            ReplaceWith("PianoId.getInstance().signIn()")
+        )
+        fun signIn(): PianoIdClient.SignInContext = getInstance().signIn()
 
         /**
          * Sign out user by it's token
@@ -96,23 +161,14 @@ class PianoId {
         @Suppress("unused") // Public API.
         @JvmStatic
         @JvmOverloads
+        @Deprecated(
+            "Use method directly from PianoIdClient instance, will be removed in future versions",
+            ReplaceWith("PianoId.getInstance().signOut(accessToken, callback)")
+        )
         fun signOut(accessToken: String, callback: PianoIdFuncCallback<Any>? = null) {
             val signOutCallback = callback ?: {}
             client?.signOut(accessToken, signOutCallback)
                 ?: signOutCallback(Result.failure(IllegalStateException(NOT_INITIALIZED_MSG)))
-        }
-
-        /**
-         * Sign out user by it's token
-         *
-         * @param accessToken User access token
-         */
-        @Suppress("unused") // Public API.
-        @JvmStatic
-        suspend fun signOut(accessToken: String) = suspendCancellableCoroutine { continuation ->
-            signOut(accessToken) { result ->
-                continuation.resumeWith(result)
-            }
         }
 
         /**
@@ -124,22 +180,13 @@ class PianoId {
         @Suppress("unused") // Public API.
         @Throws(IllegalStateException::class)
         @JvmStatic
+        @Deprecated(
+            "Use method directly from PianoIdClient instance, will be removed in future versions",
+            ReplaceWith("PianoId.getInstance().refreshToken(refreshToken, callback)")
+        )
         fun refreshToken(refreshToken: String, callback: PianoIdFuncCallback<PianoIdToken>) {
             client?.refreshToken(refreshToken, callback)
                 ?: callback(Result.failure(IllegalStateException(NOT_INITIALIZED_MSG)))
-        }
-
-        /**
-         * Refresh user access token
-         *
-         * @param refreshToken User refresh token
-         */
-        @Suppress("unused") // Public API.
-        @JvmStatic
-        suspend fun refreshToken(refreshToken: String) = suspendCancellableCoroutine { continuation ->
-            refreshToken(refreshToken) { result ->
-                continuation.resumeWith(result)
-            }
         }
 
         /**
@@ -152,6 +199,10 @@ class PianoId {
         @Suppress("unused") // Public API.
         @JvmStatic
         @JvmOverloads
+        @Deprecated(
+            "Use method directly from PianoIdClient instance, will be removed in future versions",
+            ReplaceWith("PianoId.getInstance().getUserInfo(accessToken, formName, callback)")
+        )
         fun getUserInfo(
             accessToken: String,
             formName: String? = null,
@@ -162,24 +213,6 @@ class PianoId {
         }
 
         /**
-         * Gets user info
-         *
-         * @param accessToken User access token
-         * @param formName Form name, which stores data. Use null for default
-         */
-        @Suppress("unused") // Public API.
-        @JvmStatic
-        @JvmOverloads
-        suspend fun getUserInfo(
-            accessToken: String,
-            formName: String? = null,
-        ) = suspendCancellableCoroutine { continuation ->
-            getUserInfo(accessToken, formName) { result ->
-                continuation.resumeWith(result)
-            }
-        }
-
-        /**
          * Stores user info
          *
          * @param accessToken User access token
@@ -188,6 +221,10 @@ class PianoId {
          */
         @Suppress("unused") // Public API.
         @JvmStatic
+        @Deprecated(
+            "Use method directly from PianoIdClient instance, will be removed in future versions",
+            ReplaceWith("PianoId.getInstance().putUserInfo(accessToken, newUserInfo, callback)")
+        )
         fun putUserInfo(
             accessToken: String,
             newUserInfo: PianoUserInfo,
@@ -197,28 +234,11 @@ class PianoId {
                 ?: callback(Result.failure(IllegalStateException(NOT_INITIALIZED_MSG)))
         }
 
-        /**
-         * Stores user info
-         *
-         * @param accessToken User access token
-         * @param newUserInfo New user info
-         */
-        @Suppress("unused") // Public API.
-        @JvmStatic
-        suspend fun putUserInfo(
-            accessToken: String,
-            newUserInfo: PianoUserInfo,
-        ) = suspendCancellableCoroutine { continuation ->
-            putUserInfo(accessToken, newUserInfo) { result ->
-                continuation.resumeWith(result)
-            }
-        }
-
         @Suppress("unused") // Public API.
         @JvmStatic
         fun Uri?.parsePianoIdToken(callback: PianoIdFuncCallback<PianoIdToken>) =
             if (isPianoIdUri()) {
-                getClient().parseToken(this!!, callback)
+                getInstance().parseToken(this!!, callback)
             } else {
                 callback(Result.failure(PianoIdException("It's not Piano ID uri")))
             }
@@ -238,11 +258,6 @@ class PianoId {
                 scheme?.lowercase(Locale.ENGLISH)?.startsWith(PianoIdClient.LINK_SCHEME_PREFIX) == true &&
                     PianoIdClient.LINK_AUTHORITY.equals(authority, ignoreCase = true)
             } ?: false
-
-        @JvmStatic
-        internal fun getClient() = checkNotNull(client) {
-            NOT_INITIALIZED_MSG
-        }
 
         /**
          * Default production endpoint