+++ id/id/src/test/java/android/util/Base64.kt
@@ -3,4 +3,7 @@ package android.util
 object Base64 {
     @JvmStatic
     fun decode(str: String, flags: Int): ByteArray = str.toByteArray()
+
+    @JvmStatic
+    fun encode(bytes: ByteArray, flags: Int): ByteArray = bytes
 }