+++ show-custom-form/src/main/java/io/piano/android/showform/ShowFormController.kt
@@ -5,11 +5,21 @@ import android.webkit.WebView
 import com.squareup.moshi.Moshi
 import io.piano.android.composer.model.Event
 import io.piano.android.composer.model.events.ShowForm
-import io.piano.android.id.FormUrlBuilder
+import io.piano.android.id.FormHelper
 import io.piano.android.id.PianoId
 import io.piano.android.showhelper.BaseShowController
 import timber.log.Timber
 
+/**
+ * A controller for displaying a custom form using the Piano ID API.
+ *
+ * This controller is responsible for showing a custom form specified by the [ShowForm] event.
+ * It extends the [BaseShowController] with [ShowForm] data and handles form display.
+ *
+ * @param event The [Event] containing the [ShowForm] data to be displayed.
+ * @param initialToken Initial value for user access token.
+ * @param loginCallback Callback, which be called if current user access token is invalid.
+ */
 class ShowFormController(event: Event<ShowForm>, initialToken: String = "", private val loginCallback: () -> Unit) :
     BaseShowController<ShowForm, ShowFormJs>(
         event.eventData,
@@ -19,7 +29,7 @@ class ShowFormController(event: Event<ShowForm>, initialToken: String = "", priv
     private val trackingId = event.eventExecutionContext.trackingId
     override val url: String by lazy {
         with(eventData) {
-            FormUrlBuilder.buildUrl(formName, hideCompletedFields, trackingId)
+            FormHelper.buildUrl(formName, hideCompletedFields, trackingId)
         }
     }
     override val fragmentTag = FRAGMENT_TAG
@@ -33,7 +43,7 @@ class ShowFormController(event: Event<ShowForm>, initialToken: String = "", priv
         if (accessToken.isEmpty()) {
             callback(true)
         } else {
-            PianoId.getUserInfo(accessToken, eventData.formName) { r ->
+            PianoId.getInstance().getUserInfo(accessToken, eventData.formName) { r ->
                 val shouldHideForm = r.onFailure {
                     Timber.w(it)
                 }.getOrNull()?.allCustomFieldsFilled ?: false
@@ -53,6 +63,11 @@ class ShowFormController(event: Event<ShowForm>, initialToken: String = "", priv
 
     override fun WebView.configure() = prepare(jsInterface, loginCallback = loginCallback)
 
+    /**
+     * Updates user token for form.
+     *
+     * @param userToken The updated user token
+     */
     fun updateToken(userToken: String) {
         jsInterface.token = userToken
         if (checkProfileAtTokenChange) {